Refactored Header component: Split into HeaderDesktop and HeaderMobile for better code structure and maintainability
